requirements.txt
aiofiles
aioshutil
anytree
apscheduler
aria2p
beautifulsoup4
cloudscraper
feedparser
flask
gevent
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
httpx
lxml
motor
natsort
pillow
psutil
pycryptodome
pycurl
pymongo
pyrofork
python-dotenv
python-magic
qbittorrent-api
requests
speedtest-cli
telegraph
tenacity
TgCrypto-pyrofork
urllib3
uvloop
xattr
yt-dlp